Move output.output to abc.output
