+++ logprep/connector/elasticsearch/output.py
@@ -10,7 +10,7 @@ import elasticsearch
 from elasticsearch import helpers
 from logprep.connector.connector_factory_error import InvalidConfigurationError
 from logprep.abc.input import Input
-from logprep.output.output import FatalOutputError, Output
+from logprep.abc.output import FatalOutputError, Output
 
 
 class ElasticsearchOutputFactory:
